Various wrong things

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -3,6 +3,7 @@ package com.techelevator.tenmo.dao;
 import com.techelevator.tenmo.model.Account;
 
 import javax.security.auth.login.AccountNotFoundException;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountDao {
@@ -11,9 +12,9 @@ public interface AccountDao {
 
     Account find(int accountId) throws AccountNotFoundException;
 
-    Account updateBalance(Account account, double balanceAmount) throws AccountNotFoundException;
+    Account updateBalance(Account account, BigDecimal balanceAmount) throws AccountNotFoundException;
 
-    double getBalanceByUserId(int userId) throws AccountNotFoundException;
+    BigDecimal getBalanceByUserId(int userId) throws AccountNotFoundException;
 
     void delete(int accountId);
